+++ src/epsilonServer.h
@@ -6,7 +6,7 @@
 class EpsilonServer : public GameServer
 {
 public:
-    EpsilonServer(int server_port);
+    EpsilonServer(uint16_t server_port);
     virtual ~EpsilonServer() = default;
 
     virtual void onNewClient(int32_t client_id) override;
